chore: Renames exploatory_view_config module (#121120)
