+++ x-pack/plugins/observability/public/components/shared/exploratory_view/contexts/exploratory_view_config.tsx
