+++ x-pack/plugins/observability/public/components/shared/exploratory_view/hooks/use_lens_attributes.test.tsx
@@ -18,7 +18,7 @@ import * as lensAttributes from '../configurations/lens_attributes';
 import * as indexPattern from './use_app_index_pattern';
 import * as theme from '../../../../hooks/use_theme';
 import { dataTypes, obsvReportConfigMap, reportTypesList } from '../obsv_exploratory_view';
-import { ExploratoryViewContextProvider } from '../contexts/exploatory_view_config';
+import { ExploratoryViewContextProvider } from '../contexts/exploratory_view_config';
 import { themeServiceMock } from 'src/core/public/mocks';
 
 const mockSingleSeries = [
