+++ x-pack/plugins/observability/public/components/shared/exploratory_view/rtl_helpers.tsx
@@ -48,7 +48,7 @@ import { ListItem } from '../../../hooks/use_values_list';
 import { TRANSACTION_DURATION } from './configurations/constants/elasticsearch_fieldnames';
 import { casesPluginMock } from '../../../../../cases/public/mocks';
 import { dataTypes, obsvReportConfigMap, reportTypesList } from './obsv_exploratory_view';
-import { ExploratoryViewContextProvider } from './contexts/exploatory_view_config';
+import { ExploratoryViewContextProvider } from './contexts/exploratory_view_config';
 
 interface KibanaProps {
   services?: KibanaServices;
